requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
redis==5.0.8
structlog==24.1.0
prometheus-client==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
PyYAML==6.0.2
jsonschema==4.23.0
psycopg[binary]==3.2.1

